CMS-1699: Add filter handler file
